\name{example_data}
\docType{data}
\alias{example_data}
\title{Example patient data}
\description{
  Example of a  brief list of breast cancer patient records with the necessary variables to calculate Predict v2.1 scores.
}
\usage{data(example_data)}
\format{A dataframe with 50 patient observations and 13 variables.}
\keyword{datasets}
