\name{scan_example_data}
\docType{data}
\alias{scan_example_data}
\title{Example SCAN patient data}
\description{
  Example of a  brief list of breast cancer patient records with the necessary variables to calculate Predict v2.1 scores, according to coding and naming conventions of SCAN.
}
\usage{data(scan_example_data)}
\format{A dataframe with 20 patient observations and 8 variables.}
\keyword{datasets}
